Move reading of scopes yaml file out of generator (#44)
Currently, you must provide the generator with a yaml file containing
the scope descriptions and it reads that file.  Having the generator
read from the file system feels out of place since that should really be
the plugins responsibility.

Added a new `scopes` property to `Oauth2Configuration` that will hold a
map of the scopes and descriptions.  The plugin now reads in the yaml
file and populates this new `scopes` property instead of just passing
the file to the generator.