Support operation name placeholders
Currently, if you use operation name place holders like `{class-name}`
and `{method-name}` the place holders are not replaced with the actual
values in the `operationId` property of the `resource.json` file. This
results in an invalid specification being produced since all the
operation IDs will be the same and they need to be unique.

This change fixes `ResourceSnippet` so that it properly replaces the
place holders when building the resource model.