Improvements to wrapper API
Currently, the `MockMvcRestDocumentationWrapper` provides various
overloaded `document` methods that allow you to pass in details
needed for building the resource model like `description`, `summary`,
`privateResource`, `deprecated`, etc.  This works fine but it can
be difficult to read when using all the parameters.

For instance:

```java
  document(
    "get-book-by-id",
    "Get book by id",
    true,
    false,
    responseFields(
      fieldWithPath("title").description("the title of the book"),
      fieldWithPath("author").description("the author of the book")
    )
  )
```

In that example, it's not clear that `true` refers to the
`privateResource` flag and `false` refers to the `deprecated` flag.

Another issue is that some parameters on `ResourceSnippetParameters`
are not available to be set, like `tags`.  We could add them to the
overloaded methods but that would just make the above problem worse.

This pull requests adds the ability to just provide a
`ResourceSnippetDetails` object which is a subset of
`ResourceSnippetParametersBuilder`. This allows us to change the
above example to the following:

```java
  document(
    "get-book-by-id",
    resource()
      .description("Get book by id")
      .privateResource(true)
      .deprecated(false)
      .tag("book"),
    responseFields(
      fieldWithPath("title").description("the title of the book"),
      fieldWithPath("author").description("the author of the book")
    )
  )
```