[#113] Fix for application/x-www-form-urlencoded
Main point of this fix is to introduce proper payload documentation for
OpenAPI 2.0 (changing type of parameter for a given Content-Type from
query to formData) and OpenAPI 3.0 (moving payload definition from query
parameters to requestBody).

It does not fix schema generation for OpenAPI 3.0 (it is still
generating empty schema, although - it generates proper example if
provided).