Add redoc html generation example
When I first read this readme, I thought that redoc could only serve it for me. After looking into the redoc-cli, I realized that it could just bundle me up a no-dependency HTML file. Knowing that redoc could also just produce an HTML file is something that a lot of people will likely be interested in.