Add redoc html generation example to README (#149)
* Add redoc html generation example

When I first read this readme, I thought that redoc could only serve it for me. After looking into the redoc-cli, I realized that it could just bundle me up a no-dependency HTML file. Knowing that redoc could also just produce an HTML file is something that a lot of people will likely be interested in.

* Fixes #150 by fixing bad link

I tracked down the file and it looks like it was the difference of a hyphen in the URL.

* Nix "just" per @ozscheyge suggestion

Co-authored-by: Oliver <ozscheyge@users.noreply.github.com>

* Nix other "just" per @ozscheyge suggestion

Co-authored-by: Oliver <ozscheyge@users.noreply.github.com>

Co-authored-by: Oliver <ozscheyge@users.noreply.github.com>