GH-153: Upgrade linter version
Due to version dependencies in the used plugins, we'll need to upgrade
the Kotlin version to 1.4.20 and Gradle to version 6.

This this the currently used linter version doesn't work anymore:

```

$ ./gradlew formatKotlin lintKotlin
Starting a Gradle Daemon, 2 incompatible and 5 stopped Daemons could not be reused, use --status for details
> Task :restdocs-api-spec:formatKotlinMain FAILED

FAILURE: Build failed with an exception.

* What went wrong:
Execution failed for task ':restdocs-api-spec:formatKotlinMain'.
> There was a failure while executing work items
   > A failure occurred while executing org.jmailen.gradle.kotlinter.tasks.format.FormatWorkerRunnable
      > org/jetbrains/kotlin/lexer/KtTokens.FUN_KEYWORD

* Try:
Run with --stacktrace option to get the stack trace. Run with --info or --debug option to get more log output. Run with --scan to get full insights.

```

After application of the auto fix after the version upgrade,
there will be a lot of changed lines. That's why this is coming
as a separate pull request.